requirements.txt
groq
gTTS
gradio
mysql-connector-python
python-dotenv
Pillow
opencv-python
bcrypt
geocoder